#9 (05/29) 재고 등록/수정 완료

+++ src/main/java/com/hmdandelion/project_1410002/inventory/dto/material/dto/WarehouseDTO.java
@@ -1,6 +1,6 @@
 package com.hmdandelion.project_1410002.inventory.dto.material.dto;
 
-import com.hmdandelion.project_1410002.inventory.domian.entity.product.Warehouse;
+import com.hmdandelion.project_1410002.inventory.domian.entity.warehouse.Warehouse;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 public class WarehouseDTO {
     private final String warehouseName;
     private final String location;
-    private final int volume;
+    private final long volume;
 
     public static WarehouseDTO from(Warehouse warehouse) {
         return new WarehouseDTO(
